+++ gusto/src/main/java/com/umc/gusto/global/auth/service/UserDetailsService.java
@@ -1,4 +1,4 @@
-package com.umc.gusto.global.auth;
+package com.umc.gusto.global.auth.service;
 
 import com.umc.gusto.domain.user.entity.User;
 import com.umc.gusto.domain.user.repository.UserRepository;
